Fix: css에서 스타일드 컴포넌트로 수정, log 제거
