+++ src/components/Todo.jsx
@@ -1,16 +1,13 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-
-// import CustomBtn from '../button/button';
-import './style.css';
-
+import styled from 'styled-components';
 export default function Todo(props) {
   const { todo, name, changeHandler } = props;
-  // console.log(todo.id);
+
   //체크박스가 event.target.checked가 true임에도 해제되어있어서 임의로 name이 isDone이면 defaultChecked되도록 구현함
   //렌더링이 안돼서!! 12/19 수정완료
   return (
-    <div className='Box'>
+    <Box>
       <p>
         <input
           type='checkbox'
@@ -21,10 +18,27 @@ export default function Todo(props) {
           }}
         />
         {/* Path Varialble 사용 */}
-        <Link to={`/contents/${todo.id}`} className='Link'>
+        <Link
+          to={`/contents/${todo.id}`}
+          style={{ textDecoration: 'none', color: 'black' }}
+        >
           {todo.title}
         </Link>
       </p>
-    </div>
+    </Box>
   );
 }
+
+const Box = styled.div`
+  max-width: 250px;
+  min-width: 250px;
+  border: 0px;
+  background-color: #91ab9b;
+  border-radius: 10px;
+  height: 50px;
+  align-self: center;
+  justify-content: center;
+  align-items: center;
+  margin: 10px 0;
+  display: flex;
+`;
