--- src/components/button/style.css
@@ -1,11 +1,0 @@
-.CustomBtn {
-  border-radius: 5px;
-  border: 0px;
-  background-color: #bbddc8;
-  height: 40px;
-  width: 120px;
-  min-width: 80px;
-  /* max-width: 80px; */
-  margin-left: 5px;
-  margin-right: 5px;
-}
