--- src/components/form/style.css
@@ -1,30 +1,0 @@
-form {
-  background-color: #345e44;
-  color: beige;
-  text-align: center;
-  font-size: 20px;
-  display: flex;
-  justify-content: space-between;
-  border-radius: 10px;
-}
-
-.Input_area {
-  padding-left: 20px;
-  height: 30px;
-}
-.Input_area > input {
-  width: 295px;
-  height: 30px;
-  border: 0px;
-}
-.Btn {
-  width: 250px;
-}
-
-.Btn > input {
-  width: 150px;
-  height: 35px;
-  border-radius: 10px;
-  border: 0px;
-  background-color: #bbddc8;
-}
