--- src/components/todo/style.css
@@ -1,18 +1,0 @@
-.Box {
-  max-width: 250px;
-  min-width: 250px;
-  border: 0px;
-  background-color: #91ab9b;
-  border-radius: 10px;
-  height: 50px;
-  align-self: center;
-  justify-content: center;
-  align-items: center;
-  margin: 10px 0;
-  display: flex;
-}
-
-.Link {
-  color: black;
-  text-decoration: none;
-}
